abstracting snapshot
